bring common tagging functions to a common utility file
